refactor: move event out of lib
