+++ lua/cmp/core.lua
@@ -3,6 +3,6 @@ local core = {}
 core.view = {}
 function core.view.visible() return require('blink.cmp').windows.autocomplete.win:is_open() end
 
-core.event = require('blink.compat.lib.event')
+core.event = require('blink.compat.event')
 
 return core
